Refactor dispatching code
