+++ native/src/main/java/io/ballerina/stdlib/http/api/service/signature/QueryParam.java
@@ -30,15 +30,17 @@ public class QueryParam {
     private final int typeTag;
     private final String token;
     private final boolean nilable;
+    private final boolean readonly;
     private final int index;
     private final Type type;
 
-    QueryParam(Type type, String token, int index, boolean nilable) {
+    QueryParam(Type type, String token, int index, boolean nilable, boolean readonly) {
         this.type = type;
         this.typeTag = type.getTag();
         this.token = token;
         this.index = index;
         this.nilable = nilable;
+        this.readonly = readonly;
     }
 
     public String getToken() {
@@ -59,5 +61,9 @@ public class QueryParam {
 
     public Type getType() {
         return this.type;
+    }
+
+    public boolean isReadonly() {
+        return this.readonly;
     }
 }
